Clean up ask thread history implementation
