+++ src/modules/CommandHandler.js
@@ -68,7 +68,11 @@ module.exports = class CommandHandler extends require('./Module') {
         if (!command) return;
 
         // Check if that command has been disabled in this channel
-        const channel = await this.bot.snail_db.Channel.findById(message.channel.id);
+        const disabledChannelId =
+            message.channel?.threadMetadata && message.channel?.parentID
+                ? message.channel.parentID
+                : message.channel.id;
+        const channel = await this.bot.snail_db.Channel.findById(disabledChannelId);
         if (channel?.disabledCommands.includes(command.alias[0])) {
             if (this.disabledCooldowns[message.author.id + message.command]) return;
